monitoring/observability-demo/services/order-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
requests==2.31.0
structlog==23.2.0
pydantic==2.5.0
httpx==0.25.2
